all positions in radians and velocities in radians per second (at the output of the gearbox)

+++ src/motor_control/actuator.rs
@@ -250,8 +250,8 @@ impl<'d, const N: usize> RawMotorsIO<N> for Actuator<'d, N> {
 
 
     /// Get the velocity limit of the motors (in radians per second)
-    fn get_velocity_limit(&mut self) -> Result<[u32; N]> {
-        let mut res = [0; N];
+    fn get_velocity_limit(&mut self) -> Result<[f32; N]> {
+        let mut res = [0.0; N];
         for (i, axis) in self.axes.iter_mut().enumerate() {
             res[i] = axis.get_velocity_limit()?[0];
         }
@@ -259,7 +259,7 @@ impl<'d, const N: usize> RawMotorsIO<N> for Actuator<'d, N> {
         Ok(res)
     }
     /// Set the velocity limit of the motors (in radians per second)
-    fn set_velocity_limit(&mut self, velocity: [u32; N]) -> Result<()> {
+    fn set_velocity_limit(&mut self, velocity: [f32; N]) -> Result<()> {
         for (i, axis) in self.axes.iter_mut().enumerate() {
             axis.set_velocity_limit([velocity[i]])?;
         }
