+++ src/shared_memory.rs
@@ -26,7 +26,7 @@ pub struct Memory<const N: usize> {
 
     uq_ud_limit: [i16;N],
     torque_flux_limit: [u16;N],
-    velocity_limit: [u32;N],
+    velocity_limit: [f32;N],
 
 
     axis_sensor: [f32; N],
@@ -187,10 +187,10 @@ impl<const N: usize> SharedMemory<N> {
 	self.inner.borrow_mut().torque_flux_limit=limit;
     }
 
-    pub fn get_velocity_limit(&self) -> [u32;N] {
+    pub fn get_velocity_limit(&self) -> [f32;N] {
 	self.inner.borrow().velocity_limit
     }
-    pub fn set_velocity_limit(&self, limit: [u32;N]) {
+    pub fn set_velocity_limit(&self, limit: [f32;N]) {
 	self.inner.borrow_mut().velocity_limit=limit;
     }
 
@@ -250,8 +250,8 @@ impl<const N: usize> SharedMemory<N> {
 
 		uq_ud_limit: [0;N],
 		torque_flux_limit: [0;N],
-		velocity_limit: [0;N],
-        velocity_feedforward: [0.0; N],
+		velocity_limit: [0.0;N],
+                velocity_feedforward: [0.0; N],
 
 
 		error_led: false,
@@ -284,8 +284,8 @@ impl<const N: usize> SharedMemory<N> {
 	    // velocity_limit: actuator.get_velocity_limit().unwrap_or([f32::NAN;N]),
 	    uq_ud_limit: actuator.get_uq_ud_limit().unwrap_or([0;N]),
 	    torque_flux_limit: actuator.get_torque_flux_limit().unwrap_or([0;N]),
-	    velocity_limit: actuator.get_velocity_limit().unwrap_or([0;N]),
-        velocity_feedforward: actuator.get_velocity_feedforward().unwrap_or([0.0; N]),
+	    velocity_limit: actuator.get_velocity_limit().unwrap_or([0.0;N]),
+            velocity_feedforward: actuator.get_velocity_feedforward().unwrap_or([0.0; N]),
 
 	    #[cfg(feature = "orbita3d")]
 	    index_sensor: actuator.get_index_sensor(),
